requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap5==0.7
Django==4.1
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
djongo==1.3.6
dnspython==2.3.0
Pillow==9.5.0
pymongo==3.12.3
python-dotenv==1.0.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.2.4
typing_extensions==4.6.3
tzdata==2023.3
